# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/signup": {
      "post": {
        "operationId": "AuthController_signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Signed Up",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignUpDto"
                }
              }
            }
          }
        },
        "summary": "Signing Up new User",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Logged In",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginDto"
                }
              }
            }
          }
        },
        "summary": "Login User",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/verify-login-otp": {
      "post": {
        "operationId": "AuthController_verifyOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyLoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Verify Login OTP",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/change-password": {
      "patch": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Change password (Authenticated users only)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Request password reset link",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/reset-password": {
      "patch": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Reset password using token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/roles/{id}": {
      "get": {
        "operationId": "RolesController_findRoleById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Role ID (UUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns role details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetRolesDto"
                }
              }
            }
          },
          "404": {
            "description": "Role not found"
          }
        },
        "summary": "Get role by ID",
        "tags": [
          "Roles"
        ]
      }
    },
    "/roles/updateRole/{id}": {
      "put": {
        "operationId": "RolesController_updateRoleById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Role ID (UUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Role updated successfully"
          },
          "404": {
            "description": "Role not found"
          }
        },
        "summary": "Update role by ID",
        "tags": [
          "Roles"
        ]
      }
    },
...
    "/events/create": {
      "post": {
        "operationId": "EventsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new event (Organizer only)",
        "tags": [
          "Events"
        ]
      }
    },
    "/events/allById": {
      "get": {
        "operationId": "EventsController_findAllId",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "location",
            "required": false,
            "in": "query",
            "description": "Optional filter by location",
            "schema": {
              "example": "NYC",
              "type": "string"
            }
          },
          {
            "name": "pricing",
            "required": false,
            "in": "query",
            "description": "Optional filter by pricing category",
            "schema": {
              "example": "VIP",
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Filter events by user ID",
            "schema": {
              "example": "f6aab3a8-31fe-466c-92ed-92ac5f7d96e2",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of events retrieved"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all events by logged-in user (Organizer/Admin)",
        "tags": [
          "Events"
        ]
      }
    },
    "/events/{id}": {
      "get": {
        "operationId": "EventsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Event ID (UUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get event by ID",
        "tags": [
          "Events"
        ]
      }
    },
    "/events/edit/{id}": {
      "patch": {
        "operationId": "EventsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Event ID (UUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update event (Organizer/Admin)",
        "tags": [
          "Events"
        ]
      }
    },
    "/events/delete/{id}": {
      "delete": {
        "operationId": "EventsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Event ID (UUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event deleted successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete event (Organizer/Admin Only)",
        "tags": [
          "Events"
        ]
      }
    },
...
  },
  "info": {
    "title": "Event Management Website",
    "description": "Website for online tickets booking!",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SignUpDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Full name of the user",
            "example": "Abdullah Ahmed"
          },
          "email": {
            "type": "string",
            "description": "User's email address (must be unique)",
            "example": "abdullah12@gmail.com"
          },
          "password": {
            "type": "string",
            "description": "Password should be atleast 8 digits long",
            "minLength": 8,
            "example": "StrongPass123"
          },
          "role": {
            "type": "string",
            "description": "Role to assign to the user",
            "enum": [
              "organizer",
              "customer",
              "admin"
            ],
            "example": "customer"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "role"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "abdullah12@gmail.com"
          },
          "password": {
            "type": "string",
            "description": "plaw12!@"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "VerifyLoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Enter Email: "
          },
          "otp": {
            "type": "string",
            "description": "Enter OTP: "
          }
        },
        "required": [
          "email",
          "otp"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string",
            "description": "Enter old Password: "
          },
          "newPassword": {
            "type": "string",
            "description": "Ënter New Password: "
          }
        },
        "required": [
          "oldPassword",
          "newPassword"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "abdullah12@gmail.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "abdullah12@gmail.com"
          },
          "otp": {
            "type": "string",
            "description": "232423"
          },
          "password": {
            "type": "string",
            "description": "plaw12!@"
          }
        },
        "required": [
          "email",
          "otp",
          "password"
        ]
      },
      "GetRolesDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Role Id for the user",
            "example": "f9f0b7f3-f332-4246-8460-ae7b6491e40f"
          }
        },
        "required": [
          "id"
        ]
      },
      "Permission": {
        "type": "object",
        "properties": {}
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "permissions": {
            "description": "Updated list of permissions for the role",
            "example": [
              {
                "actions": [
                  "read"
                ],
                "resource": "events"
              },
              {
                "actions": [
                  "write"
                ],
                "resource": "events"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          }
        },
        "required": [
          "permissions"
        ]
      },
...
      "CreatePricingDto": {
        "type": "object",
        "properties": {
          "tier": {
            "type": "string",
            "description": "Tier of pricing",
            "enum": [
              "student",
              "vip",
              "general admission"
            ],
            "example": "student"
          },
          "pricing": {
            "type": "number",
            "description": "Price for this tier",
            "example": 50
          },
          "maxTickets": {
            "type": "number",
            "description": "Maximum number of tickets available for this tier",
            "example": 100
          },
          "soldTickets": {
            "type": "number",
            "description": "Number of tickets already sold (optional)",
            "example": 20
          },
          "discountPercentage": {
            "type": "number",
            "description": "Discount percentage (optional)",
            "example": 10
          },
          "discountName": {
            "type": "string",
            "description": "Discount name (optional)",
            "example": "Early Bird"
          },
          "discountStartDate": {
            "format": "date-time",
            "type": "string",
            "description": "Discount start date (optional, ISO string)",
            "example": "2025-10-01T10:00:00.000Z"
          },
          "discountEndDate": {
            "format": "date-time",
            "type": "string",
            "description": "Discount end date (optional, ISO string)",
            "example": "2025-10-10T23:59:59.000Z"
          }
        },
        "required": [
          "tier",
          "pricing",
          "maxTickets"
        ]
      },
      "CreateEventDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the event",
            "example": "AI Confer 2025"
          },
          "description": {
            "type": "string",
            "description": "Description of the event",
            "example": "An exciting tech conference about AI and cloud computing."
          },
          "startDateTime": {
            "format": "date-time",
            "type": "string",
            "description": "Start date and time of the event (ISO string)",
            "example": "2025-09-15T09:00:00.000Z"
          },
          "endDateTime": {
            "format": "date-time",
            "type": "string",
            "description": "End date and time of the event (ISO string)",
            "example": "2025-09-15T18:00:00.000Z"
          },
          "location": {
            "type": "string",
            "description": "Location of the event",
            "example": "123 Main Street, Cityville"
          },
          "pricings": {
            "description": "List of pricing tiers for the event (optional)",
            "example": [
              {
                "tier": "student",
                "pricing": 50,
                "maxTickets": 100,
                "soldTickets": 10,
                "discountPercentage": 10,
                "discountName": "Early Bird",
                "discountStartDate": "2025-09-01T00:00:00.000Z",
                "discountEndDate": "2025-09-10T23:59:59.000Z"
              },
              {
                "tier": "vip",
                "pricing": 150,
                "maxTickets": 50
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreatePricingDto"
            }
          }
        },
        "required": [
          "title",
          "description",
          "startDateTime",
          "endDateTime",
          "location"
        ]
      },
      "UpdateEventDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the event",
            "example": "AI Confer 2025"
          },
          "description": {
            "type": "string",
            "description": "Description of the event",
            "example": "An exciting tech conference about AI and cloud computing."
          },
          "startDateTime": {
            "format": "date-time",
            "type": "string",
            "description": "Start date and time of the event (ISO string)",
            "example": "2025-09-15T09:00:00.000Z"
          },
          "endDateTime": {
            "format": "date-time",
            "type": "string",
            "description": "End date and time of the event (ISO string)",
            "example": "2025-09-15T18:00:00.000Z"
          },
          "location": {
            "type": "string",
            "description": "Location of the event",
            "example": "123 Main Street, Cityville"
          },
          "pricings": {
            "description": "Updated list of pricing tiers (optional)",
            "example": [
              {
                "tier": "VIP",
                "pricing": 200,
                "maxTickets": 40
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreatePricingDto"
            }
          }
        }
      },
...
    }
  }
}